update item schema

+++ src/schema/index.js
@@ -48,7 +48,7 @@ const typeDefs = gql`
     category: String!
     status: String!
     condition: String!
-    price: Int!
+    price: Float!
     quantity: Int!
     seller: Student!
     comments: [Comment]
@@ -77,7 +77,7 @@ const typeDefs = gql`
     itemDescription: String!
     category: String!
     condition: String!
-    price: Int!
+    price: Float!
     quantity: Int
     images: [String]
   }
